scanpy 1.13 plotting compat layer
