+++ pyproject.toml
@@ -280,7 +280,6 @@ addopts = [
 ]
 markers = [
   "internet: tests that require internet",
-  "spatialdata_plot: tests that render through spatialdata-plot",
   # the below markers are just there because we import from anndata.tests.helpers,
   # which means pytest sees them despite us not using them.
   "gpu: tests that require GPU",
