+++ tests/experimental/test_qc_image.py
